move view provider helper functions into `src/viewProviders/utils/`; adjust search/collapsing test structure
