+++ src/viewProviders/resources.test.ts
@@ -43,7 +43,7 @@ import {
   loadLocalResources,
   ResourceViewProvider,
 } from "./resources";
-import { SEARCH_DECORATION_URI_SCHEME } from "./search";
+import { SEARCH_DECORATION_URI_SCHEME } from "./utils/search";
 
 describe("ResourceViewProvider methods", () => {
   let provider: ResourceViewProvider;
