+++ src/viewProviders/schemas.ts
@@ -32,8 +32,8 @@ import { SchemaRegistry } from "../models/schemaRegistry";
 import { logUsage, UserEvent } from "../telemetry/events";
 import { DisposableCollection } from "../utils/disposables";
 import { RefreshableTreeViewProvider } from "./base";
-import { updateCollapsibleStateFromSearch } from "./collapsing";
-import { filterItems, itemMatchesSearch, SEARCH_DECORATION_URI_SCHEME } from "./search";
+import { updateCollapsibleStateFromSearch } from "./utils/collapsing";
+import { filterItems, itemMatchesSearch, SEARCH_DECORATION_URI_SCHEME } from "./utils/search";
 
 const logger = new Logger("viewProviders.schemas");
