+++ src/viewProviders/utils/search.ts
@@ -1,5 +1,5 @@
 import { FileDecoration, FileDecorationProvider, ThemeColor, Uri } from "vscode";
-import { ISearchable, isSearchable } from "../models/resource";
+import { ISearchable, isSearchable } from "../../models/resource";
 
 /** Check if an item matches the provided search string. */
 export function itemMatchesSearch(item: ISearchable, searchStr: string): boolean {
